# [proof of concept] async model file listing