--- a/requirements.txt
+++ b/requirements.txt
@@ -7,6 +7,7 @@ transformers>=4.28.1
 tokenizers>=0.13.3
 sentencepiece
 safetensors>=0.4.2
+aiofiles
 aiohttp
 pyyaml
 Pillow