@mcmonkey4eva 2024-09-19
`aiofiles` wasn't doing as much as I thought it was, and it's kinda wacky that it works, but new commit replaces the external dependency with a single helper class