@mcmonkey4eva 2024-09-19
`ThreadPoolExecutor` ends up a lot cleaner than trying to jank asyncio to, yknow, help asynchronously process IO :')